Week 4/Task 1/AI_Agent/requirements.txt
langchain
langchain-community
langchain-aws
langchain-text-splitters
boto3
sseclient-py
requests
python-dotenv
pypdf
faiss-cpu
langchain-huggingface
sentence-transformers
tavily-python
